+++ js/constants.js
@@ -1,0 +1,56 @@
+const GENERATOR = {
+  PHOTOS_COUNT: {
+    min: 1,
+    max: 25
+  },
+  LIKES_COUNT: {
+    min: 15,
+    max: 200
+  },
+  COMMENTS_COUNT: {
+    min: 0,
+    max: 30
+  },
+  AVATAR_ID: {
+    min: 1,
+    max: 6
+  }
+};
+
+const COMMENT_DATA = {
+  NAMES: [
+    'Михаил',
+    'София',
+    'Александр',
+    'Анна',
+    'Максим',
+    'Мария',
+    'Артем',
+    'Ева',
+    'Макр',
+    'Виктория',
+    'Лев',
+    'Полина'
+  ],
+  MESSAGES: [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+  ],
+  DESCRIPTIONS: [
+    'Не подходи ко мне, я обиделась.',
+    'Поднимаю настроение мини–фотосессией.',
+    'Фото, заряженное на позитив.',
+    'Теплые воспоминания в холодное время года.',
+    'У природы нет плохой погоды.',
+    'Добби свободен.',
+    'Мой кот',
+    'Yummy–yummy.',
+    'Наша чайная церемония.'
+  ]
+};
+
+export { GENERATOR, COMMENT_DATA };
